REF: share more ArrowEA methods
